+++ HISTORY.md
@@ -72,6 +72,8 @@
 
   * Add Hinge Embedding Loss Function (#2229).
 
+  * Add Cosine Embedding Loss Function (#2209).
+
 ### mlpack 3.2.2
 ###### 2019-11-26
   * Add `valid` and `same` padding option in `Convolution` and `Atrous